Split cli to separate per module
